requirements.txt
asgiref==3.5.2
calcure==2.6
convertdate==2.4.0
DateTime==5.2
docutils==0.20.1
hijri-converter==2.3.1
holidays==0.18
iniconfig==1.1.1
jdatetime==4.1.0
korean-lunar-calendar==0.3.1
lockfile==0.12.2
pgi==0.0.11.2
Pillow==10.0.1
psutil==5.9.5
py==1.11.0
PyMeeus==0.5.12
python-daemon==3.0.1
python-dateutil==2.8.2
pytz==2020.1
reportlab==4.0.4
six==1.16.0
tqdm==4.64.1
watchdog==3.0.0
zope.interface==6.0
